Make use of charmbracelet/log to provide colorful looging
Make use of charmbracelet/log to provide colorful looging
