+++ internal/cmd/pgxman/install.go
@@ -1,13 +1,13 @@
-//go:build !darwin
+//go:build !linux
 
 package pgxman
 
 import "github.com/spf13/cobra"
 
 func newInstallCmd() *cobra.Command {
-	return newInstallOrUpgradeCmd(false)
+	return newContainerInstallOrUpgradeCmd("pgxman", false)
 }
 
 func newUpgradeCmd() *cobra.Command {
-	return newInstallOrUpgradeCmd(true)
+	return newContainerInstallOrUpgradeCmd("pgxman", true)
 }
