--- internal/cmd/pgxman/install_darwin.go
@@ -1,13 +1,0 @@
-//go:build darwin
-
-package pgxman
-
-import "github.com/spf13/cobra"
-
-func newInstallCmd() *cobra.Command {
-	return newContainerInstallOrUpgradeCmd("pgxman", false)
-}
-
-func newUpgradeCmd() *cobra.Command {
-	return newContainerInstallOrUpgradeCmd("pgxman", true)
-}
